djangogirls/requirements.txt
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.2
drf-yasg==1.17.0
Markdown==3.1.1
psycopg2-binary==2.8.4
